requirements.txt
Django
cx-Oracle
djangorestframework
django-rest-swagger
python-decouple
django-cors-headers